refactor: 유지보수성을 위해 AuthMember를 Member와 통합
