+++ src/main/java/com/groom/orbit/member/auth/app/AuthService.java
@@ -1,4 +1,4 @@
-package com.groom.orbit.auth.app;
+package com.groom.orbit.member.auth.app;
 
 import java.time.LocalDate;
 import java.util.concurrent.TimeUnit;
@@ -10,14 +10,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.groom.orbit.ai.app.VectorService;
 import com.groom.orbit.ai.app.dto.CreateVectorDto;
-import com.groom.orbit.auth.app.dto.LoginResponseDto;
-import com.groom.orbit.auth.dao.AuthMemberRepository;
-import com.groom.orbit.auth.dao.entity.AuthMember;
 import com.groom.orbit.common.exception.CommonException;
 import com.groom.orbit.common.exception.ErrorCode;
 import com.groom.orbit.config.security.JwtTokenProvider;
-import com.groom.orbit.config.security.kakao.KakaoReissueParams;
-import com.groom.orbit.config.security.oAuth.*;
+import com.groom.orbit.member.auth.app.dto.LoginResponseDto;
+import com.groom.orbit.member.auth.app.kakao.KakaoReissueParams;
+import com.groom.orbit.member.auth.app.oAuth.*;
+import com.groom.orbit.member.member.dao.jpa.MemberRepository;
+import com.groom.orbit.member.member.dao.jpa.entity.Member;
 
 import lombok.RequiredArgsConstructor;
 
@@ -26,7 +26,7 @@ import lombok.RequiredArgsConstructor;
 @Transactional
 public class AuthService {
 
-  private final AuthMemberRepository memberRepository;
+  private final MemberRepository memberRepository;
   private final AuthTokenGenerator authTokensGenerator;
   private final RequestOAuthInfoService requestOAuthInfoService;
   private final VectorService vectorService;
@@ -38,32 +38,32 @@ public class AuthService {
 
   public LoginResponseDto login(OAuthLoginParams params) {
     OAuthInfoResponse oAuthInfoResponse = requestOAuthInfoService.request(params);
-    AuthMember authMember = findOrCreateMember(oAuthInfoResponse);
+    Member authMember = findOrCreateMember(oAuthInfoResponse);
     authMember.setLastLogin(LocalDate.now());
     return LoginResponseDto.fromLogin(
         authTokensGenerator.generate(authMember.getId()), authMember.getKakaoNickname());
   }
 
-  public AuthMember findOrCreateMember(OAuthInfoResponse oAuthInfoResponse) {
+  public Member findOrCreateMember(OAuthInfoResponse oAuthInfoResponse) {
     return memberRepository
         .findByKakaoNickname(oAuthInfoResponse.getKakaoNickname())
         .orElseGet(() -> newUser(oAuthInfoResponse));
   }
 
-  private AuthMember newUser(OAuthInfoResponse oAuthInfoResponse) {
-    AuthMember member =
-        AuthMember.builder()
+  private Member newUser(OAuthInfoResponse oAuthInfoResponse) {
+    Member member =
+        Member.builder()
             .imageUrl(oAuthInfoResponse.getKakaoImage())
             .kakaoNickname(oAuthInfoResponse.getKakaoNickname())
             .nickname(oAuthInfoResponse.getKakaoNickname())
             .build();
 
-    AuthMember savedMember = memberRepository.save(member);
+    Member savedMember = memberRepository.save(member);
     saveVector(savedMember);
     return savedMember;
   }
 
-  private void saveVector(AuthMember member) {
+  private void saveVector(Member member) {
     CreateVectorDto vectorDto = CreateVectorDto.builder().memberId(member.getId()).build();
     vectorService.save(vectorDto);
   }
@@ -78,7 +78,7 @@ public class AuthService {
       throw new CommonException(ErrorCode.INVALID_TOKEN_ERROR);
     }
 
-    AuthMember member =
+    Member member =
         memberRepository
             .findById(memberId)
             .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_MEMBER));
