+++ src/main/java/com/groom/orbit/member/member/dao/jpa/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.groom.orbit.member.member.dao.jpa;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.groom.orbit.member.member.dao.jpa.entity.Member;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+  Optional<Member> findByKakaoNickname(String kakaoNickname);
+}
